Use experimental fmt
According to the documentation this experimental mode is superior to the
current mode as it fully saves the undo history, so undo/redo doesn't
break. However it's slows (creates/deletes a file for every save) and
it's causing problems on some Vim versions.

I don't have issues with losing undo history, but this solves the issue
with constantly closing folds on save.